cloud/flask/stats/requirements.txt
bcrypt==3.2.0
entrypoints==0.3
Flask==1.1.2
Flask-Bcrypt==0.7.1
Flask-Cors==3.0.10
httplib2==0.14.0
incremental==16.10.1
pymongo==3.11.3
PyJWT==1.7.1
requests==2.22.0
